+++ src/hooks/useInput.ts
@@ -1,0 +1,13 @@
+import { ChangeEvent, useCallback, useState } from 'react';
+
+const useInput = (initialValue: string | number) => {
+  const [value, setValue] = useState(initialValue);
+  const onChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
+    setValue(event.target.value);
+  }, []);
+  const reset = useCallback(() => setValue(initialValue), [initialValue]);
+
+  return [value, onChange, reset] as const;
+};
+
+export default useInput;
